feat: add server-side logic to manage quote del pw

+++ controllers/deleteQuotePost.js
@@ -1,12 +1,23 @@
 const db = require("../db/queries");
 
+require("dotenv").config();
+
 const deleteQuotePost = async (req, res) => {
   const { quote_id } = req.params;
+  const { adminPassword } = req.body;
 
-  const getAuthorIdByQuoteId = await db.getAuthorIdByQuoteId(quote_id);
-  await db.deleteQuoteById(quote_id);
+  const author_id = await db.getAuthorIdByQuoteId(quote_id);
 
-  res.redirect(`/${getAuthorIdByQuoteId}/quotes`);
+  if (process.env.ADMINPASSWORD === String(adminPassword)) {
+    await db.deleteQuoteById(quote_id);
+
+    //Ends middleware request response cycle
+    res.redirect(`/${author_id}/quotes`);
+    //Exit js function
+    return;
+  }
+
+  res.redirect(`/${author_id}/quotes?error_type=quoteErrorDel`);
 };
 
 module.exports = deleteQuotePost;
